add process wrapper to handle logs
